backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Before the Honeymoon - Photographer Search API"
readme = "README.md"
requires-python = ">=3.11.1"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "functions-framework",
    "google-adk>=1.3.0",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
module-name = "app"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "TCH"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]
